refactor(uninstall): isolate Hermes Portable state boundaries
